fix : added Tenant User Create Response and change active status if user logged in. Also add constraint for delete users

+++ app/Http/Controllers/TenantModule/TenantUserController.php
@@ -6,6 +6,7 @@ use App\DataObjects\RequestObjects\TenantUserCreate;
 use App\DataObjects\RequestObjects\TenantUserUpdate;
 use App\Http\Controllers\Controller;
 use App\Rules\PasswordRules;
+use App\Services\TenantModule\AuthService;
 use App\Services\TenantModule\TenantUserService;
 use App\Utility\MessageCode;
 use Illuminate\Http\JsonResponse;
@@ -16,6 +17,7 @@ class TenantUserController extends Controller
 {
     public function __construct(
         private TenantUserService $tenantUserService,
+        private AuthService $authService
     ) {
     }
 
@@ -44,7 +46,6 @@ class TenantUserController extends Controller
             'phone' => ['required', 'string', 'max:30'],
             'address' => ['nullable', 'string', 'max:100'],
             'role_id' => ['nullable', 'integer'],
-            'status' => ['nullable', 'string', 'max:20'],
         ]);
 
         if ($validator->fails()) {
@@ -61,7 +62,7 @@ class TenantUserController extends Controller
             address: $validated['address'] ?? null,
             tenantId: null,
             roleId: $validated['role_id'] ?? null,
-            status: $validated['status'] ?? 'active',
+            status: 'inactive',
         ));
 
         return $this->successResponse($user->toArray(), $this->responseMessage(MessageCode::TenantUserCreated), 201);
@@ -168,6 +169,11 @@ class TenantUserController extends Controller
 
     public function destroy(string $tenantUserCode): JsonResponse
     {
+        $currentUser = $this->authService->getCurrentUser();
+        if($currentUser->code === $tenantUserCode)
+        {
+            return $this->errorResponse(message : $this->responseMessage(MessageCode::SelfDelete),statusCode : 400);
+        }
         $this->tenantUserService->delete($this->tenantUserService->resolveIdByCode($tenantUserCode));
 
         return $this->successResponse(message: $this->responseMessage(MessageCode::TenantUserDeleted));
